feat(webrtc-utils): move the offer SDP template alongside the answer
`libp2p-webrtc-utils` owns half of the SDP pair: `sdp::answer` and
`SERVER_SESSION_DESCRIPTION` live here, while the offer template sits private in
`transports/webrtc/src/tokio/sdp.rs`. The split is visible in the source --
`SERVER_SESSION_DESCRIPTION`'s doc comment opens with "See
[`CLIENT_SESSION_DESCRIPTION`]", pointing at a constant in another crate.

Move `CLIENT_SESSION_DESCRIPTION` here with its annotations and add `sdp::offer`
mirroring `sdp::answer`, so both descriptions of the libp2p WebRTC handshake are
in one place. The offer takes the fingerprint as an argument rather than
hardcoding `Fingerprint::FF`, which keeps the two functions symmetric and leaves
the "we don't verify the client certificate" decision at the call site where it
is explained.

`libp2p-webrtc`'s `sdp::offer` becomes the same thin `RTCSessionDescription`
wrapper its `sdp::answer` already was. No behaviour change: the rendered offer is
byte-for-byte what it was.
